Reverting dateFormatted attribute
