+++ client/src/Utilities.tsx
@@ -59,20 +59,4 @@ const classnames = (...args: ClassNameValue[]) => {
   return classnames(...args);
 };
 
-const formatDate = (dateString: string): string | null => {
-  try {
-    const date = new Date(dateString);
-    return date.toLocaleString("en-US", {
-      month: "long",
-      day: "numeric",
-      year: "numeric",
-      hour: "numeric",
-      minute: "numeric",
-      hour12: false,
-    });
-  } catch (e) {
-    return null;
-  }
-};
-
-export { classnames, formatDate };
+export { classnames };
